requirements.txt
Django==5.2.7
djangorestframework
django-phonenumber-field
phonenumberslite
pre-commit
psycopg2-binary
python-decouple
